🔀 :: goal entity 변경

+++ src/main/java/com/mindway/server/v2/domain/goal/util/GoalConverter.java
@@ -6,5 +6,5 @@ import com.mindway.server.v2.domain.user.entity.User;
 import java.time.LocalDate;
 
 public interface GoalConverter {
-    Goal toEntity(User user, String created_at, String ended_at, Long goal_count);
+    Goal toEntity(User user, String created_at, String ended_at, Integer goal_count);
 }
